refactor: rename rest for github infra package
