+++ src/main/java/io/pakland/mdas/githubstats/infrastructure/github/repository/TeamGitHubRepository.java
@@ -1,4 +1,4 @@
-package io.pakland.mdas.githubstats.infrastructure.rest.repository;
+package io.pakland.mdas.githubstats.infrastructure.github.repository;
 
 import io.pakland.mdas.githubstats.application.dto.TeamDTO;
 import io.pakland.mdas.githubstats.application.dto.UserDTO;
