+++ src/main/java/io/pakland/mdas/githubstats/infrastructure/github/repository/WebClientConfiguration.java
@@ -1,4 +1,4 @@
-package io.pakland.mdas.githubstats.infrastructure.rest.repository;
+package io.pakland.mdas.githubstats.infrastructure.github.repository;
 
 import lombok.NoArgsConstructor;
 import org.springframework.http.HttpHeaders;
