+++ src/main/java/io/pakland/mdas/githubstats/infrastructure/shell/controller/UserOptionController.java
@@ -14,10 +14,10 @@ import io.pakland.mdas.githubstats.domain.Organization;
 import io.pakland.mdas.githubstats.domain.Repository;
 import io.pakland.mdas.githubstats.domain.Team;
 import io.pakland.mdas.githubstats.domain.User;
-import io.pakland.mdas.githubstats.infrastructure.rest.repository.WebClientConfiguration;
-import io.pakland.mdas.githubstats.infrastructure.rest.repository.OrganizationGitHubRepository;
-import io.pakland.mdas.githubstats.infrastructure.rest.repository.RepositoryGitHubRepository;
-import io.pakland.mdas.githubstats.infrastructure.rest.repository.TeamGitHubRepository;
+import io.pakland.mdas.githubstats.infrastructure.github.repository.WebClientConfiguration;
+import io.pakland.mdas.githubstats.infrastructure.github.repository.OrganizationGitHubRepository;
+import io.pakland.mdas.githubstats.infrastructure.github.repository.RepositoryGitHubRepository;
+import io.pakland.mdas.githubstats.infrastructure.github.repository.TeamGitHubRepository;
 import io.pakland.mdas.githubstats.domain.repository.OrganizationExternalRepository;
 import io.pakland.mdas.githubstats.domain.repository.RepositoryExternalRepository;
 import io.pakland.mdas.githubstats.domain.repository.TeamExternalRepository;
