+++ src/test/java/io/pakland/mdas/githubstats/infrastructure/github/repository/TeamGitHubRepositoryTest.java
@@ -1,7 +1,7 @@
-package io.pakland.mdas.githubstats.infrastructure.rest.repository;
+package io.pakland.mdas.githubstats.infrastructure.github.repository;
 
+import io.pakland.mdas.githubstats.application.dto.TeamDTO;
 import io.pakland.mdas.githubstats.application.exceptions.HttpException;
-import io.pakland.mdas.githubstats.application.dto.UserDTO;
 import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
 import okhttp3.mockwebserver.RecordedRequest;
@@ -21,18 +21,18 @@ import java.util.List;
 import static org.junit.jupiter.api.Assertions.*;
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-class UserGitHubRepositoryTest {
+class TeamGitHubRepositoryTest {
     private MockWebServer mockWebServer;
-    private UserGitHubRepository userGitHubRepository;
-    private String teamMembersListResponse;
+    private TeamGitHubRepository teamGitHubRepository;
+    private String organizationTeamsListResponse;
 
     @BeforeAll
     void setup() throws IOException {
         this.mockWebServer = new MockWebServer();
         this.mockWebServer.start();
         WebClientConfiguration webClientConfiguration = new WebClientConfiguration(mockWebServer.url("/").toString(), "test-api-key");
-        this.userGitHubRepository = new UserGitHubRepository(webClientConfiguration);
-        this.teamMembersListResponse = new String(Files.readAllBytes(Paths.get("src/test/java/io/pakland/mdas/githubstats/infrastructure/rest/repository/responses/TeamMembers.json")));
+        this.teamGitHubRepository = new TeamGitHubRepository(webClientConfiguration);
+        this.organizationTeamsListResponse = new String(Files.readAllBytes(Paths.get("src/test/java/io/pakland/mdas/githubstats/infrastructure/rest/repository/responses/OrganizationTeams.json")));
     }
 
     @AfterAll
@@ -43,31 +43,29 @@ class UserGitHubRepositoryTest {
     @Test
     void givenValidTeamMembersRequest_shouldCallTeamMembersEndpoint() throws InterruptedException, HttpException {
         MockResponse mockResponse = new MockResponse()
-                .setBody(this.teamMembersListResponse)
+                .setBody(this.organizationTeamsListResponse)
                 .addHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE);
         mockWebServer.enqueue(mockResponse);
 
-        userGitHubRepository.fetchUsersFromTeam("github-stats-22", "gs-developers");
+        teamGitHubRepository.fetchTeamsFromOrganization("github-stats-22");
 
         RecordedRequest request = mockWebServer.takeRequest();
-        assertEquals(String.format("/orgs/%s/teams/%s/members", "github-stats-22", "gs-developers"), request.getPath());
+        assertEquals(String.format("/orgs/%s/teams", "github-stats-22"), request.getPath());
     }
 
     @Test
     void givenValidTeamId_shouldReturnTeamMembers() throws HttpException {
 
         MockResponse mockResponse = new MockResponse()
-                .setBody(this.teamMembersListResponse)
+                .setBody(this.organizationTeamsListResponse)
                 .addHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE);
         mockWebServer.enqueue(mockResponse);
 
-        List<UserDTO> response = userGitHubRepository.fetchUsersFromTeam("github-stats-22", "gs-developers");
-        List<UserDTO> expected = new ArrayList<>();
-        expected.add(0, new UserDTO(33031570, "manerow", "https://api.github.com/users/manerow/orgs"));
-        expected.add(1, new UserDTO(48334745, "mikededo", "https://api.github.com/users/mikededo/orgs"));
-        expected.add(2, new UserDTO(54351560, "sdomingobasora", "https://api.github.com/users/sdomingobasora/orgs"));
+        List<TeamDTO> response = teamGitHubRepository.fetchTeamsFromOrganization("github-stats-22");
+        List<TeamDTO> expected = new ArrayList<>();
+        expected.add(0, new TeamDTO(7098104, "gs-developers", "gs-developers"));
 
-        assertEquals(3, response.size());
+        assertEquals(response.size(), 1);
         assertArrayEquals(response.toArray(), expected.toArray());
     }
 }
